Enhance author section with links and title updates
In the 'Authors' section of the website, we've made several updates to improve user experience and align content titles with their respective sections. By wrapping the author images in hyperlinks, we're now offering direct access to their profiles, fostering greater interaction between authors and users. This change caters to the growing need for connectivity and ease of access to authors' professional information or portfolios. Additionally, we've standardized the section titles for clarity and consistency across the site: renaming the 'Multiple Authors' title simply to 'Authors' and correcting the 'About' section title to more precisely reflect its content.

These enhancements are aimed at not only improving navigation but also at enriching the user's journey throughout the site, making it more engaging and informative.